LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scroll-markers/targeted-scroll-marker-selection.tentative-expected.txt

0
1
2
3
4
5
6
7
8
9
10
11
12
13
14
15

FAIL scroll-marker of target (idx 1) of scrollIntoView is selected assert_equals: marker 15 should be unselected. expected "rgb(255, 0, 0)" but got ""
FAIL scroll-marker of target (idx 14) of scrollIntoView is selected assert_equals: marker 15 should be unselected. expected "rgb(255, 0, 0)" but got ""
FAIL scroll-marker of parent of target of scrollIntoView is selected assert_equals: marker 15 should be unselected. expected "rgb(255, 0, 0)" but got ""
FAIL scroll-marker of earlier sibling of target of scrollIntoView is selected assert_equals: marker 15 should be unselected. expected "rgb(255, 0, 0)" but got ""
